ref: add `conversations` dir, `main.py`, `utils_{tg,pd,ask}.py`
