+++ src/other_commands.py
@@ -2,9 +2,9 @@ from telegram import Update
 from telegram.ext import ContextTypes, Application
 
 from src.tables.answer import AnswerType
-from src.user_data import UserData, UserDBCache
-from src.utils import handler_decorator, USER_DATA_KEY
-from src.utils_send import send_entity_answers_df
+from src.user_data import UserData
+from src.utils_tg import USER_DATA_KEY, handler_decorator
+from src.conversations.utils_ask import send_entity_answers_df
 
 
 @handler_decorator
